// src/app/main/logistics/litigation-file/shared/litigation-file-row.interface.ts
export interface LitigationFileSearchRow {
  litigationFile?: string;
  category?: string;
  litigationType?: string;
  typeOfElement?: string;
  isBlocking?: boolean;
  folderNumber?: string;
  documentType?: string;
  status?: string;
  createdOn?: string;
  foundOn?: Date;
  createdBy?: string;
  comment?: string;
  supplier?: string;
  carrier?: string;
  criticity?: string;
  folderName?: string;
  origin?: string;
}